systemd: Move graphs code to a separate file and stop showing them in the main page
Instead show them now in $cockpit_url:/system/graphs

This is not accessible by any means other than the URL itself and it
is expected that a follow up commit will introduce overview page
that will make the graphs URL accessible through a link.
