+++ pkg/systemd/host.js
@@ -28,11 +28,11 @@ import cockpit from "cockpit";
 import * as machine_info from "machine-info.js";
 import * as packagekit from "packagekit.js";
 import { install_dialog } from "cockpit-components-install-dialog.jsx";
-import * as plot from "plot.js";
 import * as service from "service.js";
 import { shutdown } from "./shutdown.js";
 import host_keys_script from "raw-loader!./ssh-list-host-keys.sh";
 import { page_status } from "notifications";
+import { set_page_link, dialog_setup, page_show } from './helpers.js';
 
 import { PageStatusNotifications } from "./page-status.jsx";
 
@@ -44,7 +44,6 @@ import "bootstrap-datepicker/dist/js/bootstrap-datepicker";
 import "patternfly-bootstrap-combobox/js/bootstrap-combobox";
 
 const _ = cockpit.gettext;
-var C_ = cockpit.gettext;
 
 var permission = cockpit.permission({ admin: true });
 $(permission).on("changed", update_hostname_privileged);
@@ -235,16 +234,6 @@ PageServer.prototype = {
         return null;
     },
 
-    relayout: function() {
-        var self = this;
-        if (self.cpu_plot) {
-            self.cpu_plot.resize();
-            self.memory_plot.resize();
-            self.network_plot.resize();
-            self.disk_plot.resize();
-        }
-    },
-
     setup: function() {
         var self = this;
         update_hostname_privileged();
@@ -382,8 +371,6 @@ PageServer.prototype = {
                          cockpit.transport.host);
         });
 
-        self.plot_controls = plot.setup_plot_controls($('#server'), $('#server-graph-toolbar'));
-
         var pmcd_service = service.proxy("pmcd");
         var pmlogger_service = service.proxy("pmlogger");
         var pmlogger_promise;
@@ -470,19 +457,6 @@ PageServer.prototype = {
         $(pmlogger_service).on('changed', pmlogger_service_changed);
         pmlogger_service_changed();
 
-        // if cockpit component "page" is available, set element content to a link to it, otherwise just text
-        function set_page_link(element_sel, page, text) {
-            if (cockpit.manifests[page]) {
-                var link = document.createElement("a");
-                link.innerHTML = text;
-                link.tabIndex = 0;
-                link.addEventListener("click", function() { cockpit.jump("/" + page) });
-                $(element_sel).html(link);
-            } else {
-                $(element_sel).text(text);
-            }
-        }
-
         function refresh_os_updates_state() {
             const status = page_status.get("updates") || { };
             const details = status.details || { };
@@ -530,7 +504,6 @@ PageServer.prototype = {
         machine_id.read()
                 .done(function(content) {
                     $("#system_machine_id").text(content);
-                    self.relayout();
                 })
                 .fail(function(ex) {
                     console.error("Error reading machine id", ex);
@@ -555,149 +528,6 @@ PageServer.prototype = {
                                                 '/org/freedesktop/hostname1');
         self.kernel_hostname = null;
 
-        /* CPU graph */
-
-        var cpu_data = {
-            direct: ["kernel.all.cpu.nice", "kernel.all.cpu.user", "kernel.all.cpu.sys"],
-            internal: ["cpu.basic.nice", "cpu.basic.user", "cpu.basic.system"],
-            units: "millisec",
-            derive: "rate",
-            factor: 0.1 // millisec / sec -> percent
-        };
-
-        var cpu_options = plot.plot_simple_template();
-        $.extend(cpu_options.yaxis,
-                 {
-                     tickFormatter: function(v) { return v.toFixed(0) },
-                     max: 100
-                 }
-        );
-        self.cpu_plot = new plot.Plot($("#server_cpu_graph"), 300);
-        self.cpu_plot.set_options(cpu_options);
-        // This is added to the plot once we have the machine info, see below.
-
-        /* Memory graph */
-
-        var memory_data = {
-            direct: ["mem.physmem", "mem.util.available"],
-            internal: ["memory.used"],
-            units: "bytes"
-        };
-
-        var memory_options = plot.plot_simple_template();
-        $.extend(memory_options.yaxis,
-                 {
-                     ticks: plot.memory_ticks,
-                     tickFormatter: plot.format_bytes_tick_no_unit
-                 });
-        memory_options.post_hook = function memory_post_hook(pl) {
-            var axes = pl.getAxes();
-            $('#server_memory_unit').text(plot.bytes_tick_unit(axes.yaxis));
-        };
-
-        self.memory_plot = new plot.Plot($("#server_memory_graph"), 300);
-        self.memory_plot.set_options(memory_options);
-        self.memory_plot.add_metrics_difference_series(memory_data, { });
-
-        /* Network graph */
-
-        var network_data = {
-            direct: ["network.interface.total.bytes"],
-            internal: ["network.interface.tx", "network.interface.rx"],
-            "omit-instances": ["lo"],
-            units: "bytes",
-            derive: "rate"
-        };
-
-        var network_options = plot.plot_simple_template();
-        $.extend(network_options.yaxis, { tickFormatter: plot.format_bits_per_sec_tick_no_unit });
-        network_options.setup_hook = function network_setup_hook(pl) {
-            var axes = pl.getAxes();
-            if (axes.yaxis.datamax < 100000)
-                axes.yaxis.options.max = 100000;
-            else
-                axes.yaxis.options.max = null;
-            axes.yaxis.options.min = 0;
-        };
-        network_options.post_hook = function network_post_hook(pl) {
-            var axes = pl.getAxes();
-            $('#server_network_traffic_unit').text(plot.bits_per_sec_tick_unit(axes.yaxis));
-        };
-
-        self.network_plot = new plot.Plot($("#server_network_traffic_graph"), 300);
-        self.network_plot.set_options(network_options);
-        self.network_plot.add_metrics_sum_series(network_data, { });
-
-        /* Disk IO graph */
-
-        var disk_data = {
-            direct: ["disk.all.total_bytes"],
-            internal: ["disk.all.read", "disk.all.written"],
-            units: "bytes",
-            derive: "rate"
-        };
-
-        var disk_options = plot.plot_simple_template();
-        $.extend(disk_options.yaxis,
-                 {
-                     ticks: plot.memory_ticks,
-                     tickFormatter: plot.format_bytes_per_sec_tick_no_unit
-                 });
-        disk_options.setup_hook = function disk_setup_hook(pl) {
-            var axes = pl.getAxes();
-            if (axes.yaxis.datamax < 100000)
-                axes.yaxis.options.max = 100000;
-            else
-                axes.yaxis.options.max = null;
-            axes.yaxis.options.min = 0;
-        };
-        disk_options.post_hook = function disk_post_hook(pl) {
-            var axes = pl.getAxes();
-            $('#server_disk_io_unit').text(plot.bytes_per_sec_tick_unit(axes.yaxis));
-        };
-
-        self.disk_plot = new plot.Plot($("#server_disk_io_graph"), 300);
-        self.disk_plot.set_options(disk_options);
-        self.disk_plot.add_metrics_sum_series(disk_data, { });
-
-        machine_info.cpu_ram_info()
-                .done(function(info) {
-                    $('#link-cpu').text(cockpit.format(cockpit.ngettext("of $0 CPU core", "of $0 CPU cores",
-                                                                        info.cpus),
-                                                       info.cpus));
-                    cpu_data.factor = 0.1 / info.cpus; // millisec / sec -> percent
-                    self.cpu_plot.add_metrics_sum_series(cpu_data, { });
-
-                    if (info.swap) {
-                        memory_options.yaxis.max = info.memory + info.swap * 0.25;
-                        memory_options.yaxis.tickFormatter = function(v) {
-                            return v <= info.memory ? plot.format_bytes_tick_no_unit(v, memory_options.yaxis)
-                                : plot.format_bytes_tick_no_unit(v + (v - info.memory) * 4, memory_options.yaxis);
-                        };
-                        memory_options.colors[1] = "#CC0000";
-                        memory_options.grid.markings = [
-                            { yaxis: { from: info.memory, to: info.memory + info.swap * 0.25 }, color: "#ededed" }
-                        ];
-                        var swap_data = {
-                            internal: ["memory.swap-used"],
-                            units: "bytes",
-                            factor: 0.25,
-                            threshold: info.memory,
-                            offset: info.memory
-                        };
-                        self.memory_plot.add_metrics_sum_series(swap_data, { });
-                        $("#link-memory").hide();
-                        $("#link-memory-and-swap").show();
-                    } else {
-                        memory_options.yaxis.max = info.memory;
-                    }
-                    self.memory_plot.set_options(memory_options);
-                });
-
-        self.plot_controls.reset([self.cpu_plot, self.memory_plot, self.network_plot, self.disk_plot]);
-
-        $(window).on('resize.server', () => { self.relayout() });
-
         const asset_tag_text = $("#system_information_asset_tag_text");
         const hardware_text = $("#system_information_hardware_text");
         hardware_text.tooltip({ title: _("Click to see system hardware information"), placement: "bottom" });
@@ -717,7 +547,6 @@ PageServer.prototype = {
                     asset_tag_text.text(fields.product_serial || fields.chassis_serial);
                     asset_tag_text.toggle(present);
                     asset_tag_text.prev().toggle(present);
-                    self.relayout();
                 })
                 .fail(function(ex) {
                     debug("couldn't read dmi info: " + ex);
@@ -764,19 +593,10 @@ PageServer.prototype = {
     },
 
     show: function() {
-        this.cpu_plot.start_walking();
-        this.memory_plot.start_walking();
-        this.disk_plot.start_walking();
-        this.network_plot.start_walking();
     },
 
     leave: function() {
         var self = this;
-
-        self.cpu_plot.destroy();
-        self.memory_plot.destroy();
-        self.disk_plot.destroy();
-        self.network_plot.destroy();
 
         $(self.hostname_proxy).off();
         self.hostname_proxy = null;
@@ -1521,223 +1341,6 @@ function PageSystemInformationChangeSystime() {
     this._init();
 }
 
-PageCpuStatus.prototype = {
-    _init: function() {
-        this.id = "cpu_status";
-    },
-
-    getTitle: function() {
-        return C_("page-title", "CPU Status");
-    },
-
-    enter: function() {
-        var self = this;
-
-        var n_cpus = 1;
-
-        var options = {
-            series: {
-                shadowSize: 0,
-                lines: { lineWidth: 0, fill: 1 }
-            },
-            yaxis: {
-                min: 0,
-                max: n_cpus * 1000,
-                show: true,
-                ticks: 5,
-                tickFormatter: function(v) { return (v / 10 / n_cpus) + "%" }
-            },
-            xaxis: {
-                show: true,
-                ticks: [[0.0 * 60, "5 min"],
-                    [1.0 * 60, "4 min"],
-                    [2.0 * 60, "3 min"],
-                    [3.0 * 60, "2 min"],
-                    [4.0 * 60, "1 min"]]
-            },
-            x_rh_stack_graphs: true
-        };
-
-        var metrics = [
-            { name: "cpu.basic.iowait", derive: "rate" },
-            { name: "cpu.basic.system", derive: "rate" },
-            { name: "cpu.basic.user", derive: "rate" },
-            { name: "cpu.basic.nice", derive: "rate" },
-        ];
-
-        var series = [
-            { color: "#cc0000", label: _("I/O Wait") },
-            { color: "#f5c12e", label: _("Kernel") },
-            { color: "#8461f7", label: _("User") },
-            { color: "#6eb664", label: _("Nice") },
-        ];
-
-        self.channel = cockpit.metrics(1000, {
-            source: "internal",
-            metrics: metrics,
-            cache: "cpu-status-rate"
-        });
-
-        /* The grid shows us the last five minutes */
-        self.grid = cockpit.grid(1000, -300, -0);
-
-        var i;
-        for (i = 0; i < series.length; i++) {
-            series[i].row = self.grid.add(self.channel, [metrics[i].name]);
-        }
-
-        /* Start pulling data, and make the grid follow the data */
-        self.channel.follow();
-        self.grid.walk();
-
-        this.plot = plot.setup_complicated_plot("#cpu_status_graph", self.grid, series, options);
-
-        machine_info.cpu_ram_info()
-                .done(function(info) {
-                    // Setting n_cpus changes the tick labels, see tickFormatter above.
-                    n_cpus = info.cpus;
-                    self.plot.set_yaxis_max(n_cpus * 1000);
-                    $("#cpu_status_title").text(cockpit.format(cockpit.ngettext("Usage of $0 CPU core",
-                                                                                "Usage of $0 CPU cores",
-                                                                                n_cpus),
-                                                               n_cpus));
-                });
-    },
-
-    show: function() {
-        this.plot.start();
-    },
-
-    leave: function() {
-        this.plot.destroy();
-        this.channel.close();
-        this.channel = null;
-    }
-};
-
-function PageCpuStatus() {
-    this._init();
-}
-
-PageMemoryStatus.prototype = {
-    _init: function() {
-        this.id = "memory_status";
-    },
-
-    getTitle: function() {
-        return C_("page-title", "Memory");
-    },
-
-    enter: function() {
-        var self = this;
-        var dfd = cockpit.defer();
-        self.setupPromise = dfd.promise;
-
-        machine_info.cpu_ram_info().done(function(info) {
-            var options = {
-                series: {
-                    shadowSize: 0, // drawing is faster without shadows
-                    lines: { lineWidth: 0.0, fill: 1 }
-                },
-                yaxis: {
-                    min: 0,
-                    max: info.memory,
-                    ticks: 5,
-                    tickFormatter: function(v) {
-                        return cockpit.format_bytes(v);
-                    }
-                },
-                xaxis: {
-                    show: true,
-                    ticks: [[0.0 * 60, _("5 min")],
-                        [1.0 * 60, _("4 min")],
-                        [2.0 * 60, _("3 min")],
-                        [3.0 * 60, _("2 min")],
-                        [4.0 * 60, _("1 min")]]
-                },
-                x_rh_stack_graphs: true,
-            };
-            var metrics = [
-                { name: "memory.used" },
-                { name: "memory.cached" },
-            ];
-            var series = [
-                { color: "#0088ce", label: _("Used") },
-                { color: "#e4f5bc", label: _("Cached") },
-            ];
-
-            if (info.swap) {
-                options.yaxis.max = info.memory + info.swap * 0.25;
-                options.yaxis.tickFormatter = function(v) {
-                    return v <= info.memory ? cockpit.format_bytes(v)
-                        : cockpit.format_bytes(v + (v - info.memory) * 4);
-                };
-                $.extend(options, {
-                    grid: {
-                        aboveData: false, markings: [
-                            { yaxis: { from: info.memory, to: info.memory + info.swap * 0.25 }, color: "#ededed" }
-                        ]
-                    }
-                });
-                metrics.push({ name: "memory.swap-used" });
-                series.push({ color: "#e41a1c", label: _("Swap Used"), offset: info.memory, factor: 0.25 });
-            } else {
-                $("#memory_status .memory-swap").hide();
-            }
-
-            self.channel = cockpit.metrics(1000, {
-                source: "internal",
-                metrics: metrics,
-                cache: "memory-status"
-            });
-            /* The grid shows us the last five minutes */
-            self.grid = cockpit.grid(1000, -300, -0);
-            for (var i = 0; i < series.length; i++)
-                series[i].row = self.grid.add(self.channel, [metrics[i].name]);
-
-            /* Start pulling data, and make the grid follow the data */
-            self.channel.follow();
-            self.grid.walk();
-            self.plot = plot.setup_complicated_plot("#memory_status_graph", self.grid, series, options);
-            dfd.resolve();
-        })
-                .fail(function(ex) {
-                    debug("Couldn't read memory info: " + ex);
-                    dfd.reject();
-                });
-    },
-
-    show: function() {
-        var self = this;
-        if (self.setupPromise) {
-            self.setupPromise.done(function() {
-                self.plot.start();
-                $("#memory_status_graph canvas").css("z-index", -1);
-            });
-        }
-    },
-
-    leave: function() {
-        this.plot.destroy();
-        this.channel.close();
-        this.channel = null;
-    }
-};
-
-function PageMemoryStatus() {
-    this._init();
-}
-
-$("#link-cpu").on("click", function() {
-    cockpit.location.go(["cpu"]);
-    return false;
-});
-
-$("#link-memory, #link-memory-and-swap").on("click", function() {
-    cockpit.location.go(["memory"]);
-    return false;
-});
-
 $("#system_information_hardware_text").on("click", function() {
     $("#system_information_hardware_text").tooltip("hide");
     cockpit.jump("/system/hwinfo", cockpit.transport.host);
@@ -1748,89 +1351,19 @@ $("#system-information-enable-pcp-link").on("click", function() {
     install_dialog("cockpit-pcp");
 });
 
-/*
- * INITIALIZATION AND NAVIGATION
- *
- * The code above still uses the legacy 'Page' abstraction for both
- * pages and dialogs, and expects page.setup, page.enter, page.show,
- * and page.leave to be called at the right times.
- *
- * We cater to this with a little compatability shim consisting of
- * 'dialog_setup', 'page_show', and 'page_hide'.
- */
-
-function dialog_setup(d) {
-    d.setup();
-    $('#' + d.id)
-            .on('show.bs.modal', function(event) {
-                if (event.target.id === d.id)
-                    d.enter();
-            })
-            .on('shown.bs.modal', function(event) {
-                if (event.target.id === d.id)
-                    d.show();
-            })
-            .on('hidden.bs.modal', function(event) {
-                if (event.target.id === d.id)
-                    d.leave();
-            });
-}
-
-function page_show(p, arg) {
-    if (!p._entered_)
-        p.enter(arg);
-    p._entered_ = true;
-    $('#' + p.id)
-            .show()
-            .removeAttr("hidden");
-    p.show();
-}
-
-function page_hide(p) {
-    $('#' + p.id).hide();
-}
-
 function init() {
     var server_page;
-    var memory_page;
-    var cpu_page;
-
-    function navigate() {
-        var path = cockpit.location.path;
-
-        if (path.length === 0) {
-            page_hide(cpu_page);
-            page_hide(memory_page);
-            page_show(server_page);
-        } else if (path.length === 1 && path[0] == 'cpu') {
-            page_hide(server_page);
-            page_hide(memory_page);
-            page_show(cpu_page);
-        } else if (path.length === 1 && path[0] == 'memory') {
-            page_hide(server_page);
-            page_hide(cpu_page);
-            page_show(memory_page);
-        } else { /* redirect */
-            console.warn("not a system location: " + path);
-            cockpit.location = '';
-        }
-
-        $("body").removeAttr("hidden");
-    }
 
     cockpit.translate();
 
     server_page = new PageServer();
     server_page.setup();
 
-    cpu_page = new PageCpuStatus();
-    memory_page = new PageMemoryStatus();
-
     dialog_setup(new PageSystemInformationChangeHostname());
     dialog_setup(change_systime_dialog = new PageSystemInformationChangeSystime());
 
-    $(cockpit).on("locationchanged", navigate);
-    navigate();
+    page_show(server_page);
+    $("body").removeAttr("hidden");
 }
 
 $(init);
