+++ pkg/systemd/index.html
@@ -180,102 +180,6 @@
             </form>
         </div>
 
-        <div id="server-graph-columns" class="server-graphs" role="presentation">
-            <div id="server-graph-toolbar" class="zoom-controls standard-zoom-controls">
-                <div class="dropdown">
-                    <button class="btn btn-default dropdown-toggle" data-toggle="dropdown">
-                        <span></span>
-                        <i class="fa fa-caret-down pf-c-context-selector__toggle-icon" aria-hidden="true"></i>
-                    </button>
-                    <ul class="dropdown-menu" role="menu">
-                        <li role="presentation"><a tabindex="0" role="menuitem"
-                            data-action="goto-now" translate>Go to now</a></li>
-                        <li role="presentation" class="divider"></li>
-                        <li role="presentation"><a tabindex="0" role="menuitem"
-                            data-range="300" translate>5 minutes</a></li>
-                        <li role="presentation"><a tabindex="0" role="menuitem"
-                            data-range="3600" translate>1 hour</a></li>
-                        <li role="presentation"><a tabindex="0" role="menuitem"
-                            data-range="21600" translate>6 hours</a></li>
-                        <li role="presentation"><a tabindex="0" role="menuitem"
-                            data-range="86400" translate>1 day</a></li>
-                        <li role="presentation"><a tabindex="0" role="menuitem"
-                            data-range="604800" translate>1 week</a></li>
-                    </ul>
-                </div>
-                <button class="btn btn-default" data-action="zoom-out">
-                    <span class="glyphicon glyphicon-zoom-out"></span>
-                </button>
-                <div class="btn-group">
-                    <button class="btn btn-default fa fa-angle-left" data-action="scroll-left"></button>
-                    <button class="btn btn-default fa fa-angle-right" data-action="scroll-right"></button>
-                </div>
-            </div>
-            <div>
-              <span class="plot-unit" id="server_cpu_unit">%</span><a tabindex="0" id="link-cpu"></a>
-            </div>
-            <div id="server_cpu_graph" class="server-graph zoomable-plot"></div>
-            <br/>
-            <div>
-              <span class="plot-unit" id="server_memory_unit"></span>
-              <a tabindex="0" id="link-memory" translate>Memory</a>
-              <a tabindex="0" id="link-memory-and-swap" hidden translate>Memory & Swap</a>
-            </div>
-            <div id="server_memory_graph" class="server-graph zoomable-plot"></div>
-            <br/>
-            <div>
-              <span class="plot-unit" id="server_disk_io_unit"></span><span id="link-disk"></span>
-            </div>
-            <div id="server_disk_io_graph" class="server-graph zoomable-plot"></div>
-            <br/>
-            <div>
-              <span class="plot-unit" id="server_network_traffic_unit"></span><span id="link-network"></span>
-            </div>
-            <div id="server_network_traffic_graph" class="server-graph zoomable-plot"></div>
-        </div>
-    </div>
-
-    <div class="server-graph-container" id="cpu_status" hidden>
-        <div class="server-graph-title" id="cpu_status_title">&nbsp;</div>
-        <ul class="server-graph-legend">
-            <li class="cpu-nice">
-                <i class="fa fa-square"></i>
-                <span translate="yes">Nice</span>
-            </li>
-            <li class="cpu-user">
-                <i class="fa fa-square"></i>
-                <span translate="yes">User</span>
-            </li>
-            <li class="cpu-kernel">
-                <i class="fa fa-square"></i>
-                <span translate="yes">Kernel</span>
-            </li>
-            <li class="cpu-io-wait">
-                <i class="fa fa-square"></i>
-                <span translate="yes">I/O Wait</span>
-            </li>
-        </ul>
-        <div class="server-graph-graph" id="cpu_status_graph"></div>
-    </div>
-
-    <div class="server-graph-container" id="memory_status" hidden>
-        <ul class="server-graph-legend">
-            <li class="memory-swap">
-                <i class="fa fa-square"></i>
-                <span translate="yes">Swap Used</span>
-            </li>
-            <li class="memory-cached">
-                <i class="fa fa-square"></i>
-                <span translate="yes">Cached</span>
-            </li>
-            <li class="memory-used">
-                <i class="fa fa-square"></i>
-                <span translate="yes">Used</span>
-            </li>
-        </ul>
-        <div class="server-graph-graph" id="memory_status_graph"></div>
-    </div>
-
     <div class="modal" id="system_information_ssh_keys" tabindex="-1"
         role="dialog" data-backdrop="static">
         <div class="modal-dialog">
