+++ pkg/systemd/manifest.json.in
@@ -17,6 +17,10 @@
                 {
                     "matches": ["hardware", "mitigation", "pci", "memory", "cpu", "bios", "ram", "dimm", "serial"],
                     "goto": "/system/hwinfo"
+                },
+                {
+                    "matches": ["graphs"],
+                    "goto": "/system/graphs"
                 }
             ]
         },
