+++ test/verify/check-metrics
@@ -173,7 +173,10 @@ class TestMetrics(MachineCase):
     def testPcp(self):
         m = self.machine
 
-        self.login_and_go("/system")
+        if m.image in ["rhel-8-1-distropkg"]:
+            self.login_and_go("/system")
+        else:
+            self.login_and_go("/system/graphs")
 
         self.check_host_metrics()
         m.execute("pgrep cockpit-pcp")
@@ -184,7 +187,10 @@ class TestMetrics(MachineCase):
         m.execute("! [ -f /usr/libexec/cockpit-pcp ] || rm /usr/libexec/cockpit-pcp")
         m.execute("! [ -f /usr/lib/cockpit/cockpit-pcp ] || rm /usr/lib/cockpit/cockpit-pcp")
 
-        self.login_and_go("/system")
+        if m.image in ["rhel-8-1-distropkg"]:
+            self.login_and_go("/system")
+        else:
+            self.login_and_go("/system/graphs")
 
         self.check_host_metrics()
         m.execute("! pgrep cockpit-pcp")
